requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
python-multipart==0.0.12
tenacity==9.0.0
alembic==1.13.3
